prototype/idmt/pythonExampleProject/requirements.txt
paho-mqtt == 1.6.1
pyecoregen == 0.5.0
configparser == 7.1.0
flask ==  3.0.3
flask_mqtt == 1.2.1